refactor: PostManagerPage API 통신 함수 features/comment-management 변환
