+++ src/features/post-management/list/api/index.ts
@@ -1,4 +1,3 @@
-export * from "./deletePostApi"
 export * from "./filterPostApi"
 export * from "./searchPostApi"
 export * from "./compositePostApi"
