+++ src/features/post-management/update/api/index.ts
@@ -1,1 +1,2 @@
 export * from "./updatePostApi"
+export * from "./deletePostApi"
